added option to return changes as polygon
