+++ autocnet/cg/cg.py
@@ -526,3 +526,105 @@ def distribute_points_in_geom(geom, method="classic",
     else:
         print('WTF Willy')
     return valid
+
+
+def alpha_shape(points, alpha):
+    """
+    Compute a convex hull from a set of points. 
+
+    credit: https://gist.github.com/dwyerk/10561690
+
+    Parameters
+    ----------
+    
+    points : np.array
+             points in the format [[x0,y0], [x1, y,1], ... [xn, yn]]
+    
+    alpha : float 
+            Higher alphas creater a tighter the boundery around input points, alphas approaching 0 create a convex hull. 
+            Best to keep in the range (0, 1]
+
+
+    Returns 
+    -------
+
+    convex_hull : shapely.geometry.Polygon 
+                  Shapely polygon of the convex hull 
+
+    """
+    tri = Delaunay(points)
+    triangles = points[tri.vertices]
+    a = ((triangles[:,0,0] - triangles[:,1,0]) ** 2 + (triangles[:,0,1] - triangles[:,1,1]) ** 2) ** 0.5
+    b = ((triangles[:,1,0] - triangles[:,2,0]) ** 2 + (triangles[:,1,1] - triangles[:,2,1]) ** 2) ** 0.5
+    c = ((triangles[:,2,0] - triangles[:,0,0]) ** 2 + (triangles[:,2,1] - triangles[:,0,1]) ** 2) ** 0.5
+    s = ( a + b + c ) / 2.0
+    areas = (s*(s-a)*(s-b)*(s-c)) ** 0.5
+    circums = a * b * c / (4.0 * areas)
+    
+    # avoid devide by zero
+    thresh = 1.0/alpha if alpha is not 0 else circums.max()
+    filtered = triangles[circums < thresh]
+    
+    edge1 = filtered[:,(0,1)]
+    edge2 = filtered[:,(1,2)]
+    edge3 = filtered[:,(2,0)]
+    edge_points = np.unique(np.concatenate((edge1,edge2,edge3)), axis = 0).tolist()
+    m = geometry.MultiLineString(edge_points)
+    triangles = list(polygonize(m))
+    return cascaded_union(triangles)
+
+
+def rasterize_polygon(shape, vertices, dtype=bool):
+    """
+    Simple tool to convert poly into a boolean numpy array.
+    
+    source: https://stackoverflow.com/questions/37117878/generating-a-filled-polygon-inside-a-numpy-array
+    
+    Parameters
+    ----------
+    
+    shape : tuple 
+            size of the array in (y,x) format
+    
+    vertices : np.array, list
+               array of vertices in [[x0, y0], [x1, y1]...] format
+    
+    dtype : type
+            datatype of output mask
+    
+    Returns
+    -------
+    
+    mask : np.array
+           mask with filled polygon set to true
+    
+    """
+    def check(p1, p2, base_array):
+        idxs = np.indices(base_array.shape) # Create 3D array of indices
+
+        p1 = p1.astype(float)
+        p2 = p2.astype(float)
+
+        # Calculate max column idx for each row idx based on interpolated line between two points
+        if p1[0] == p2[0]:
+            max_col_idx = (idxs[0] - p1[0]) * idxs.shape[1]
+            sign = np.sign(p2[1] - p1[1])
+        else:
+            max_col_idx = (idxs[0] - p1[0]) / (p2[0] - p1[0]) * (p2[1] - p1[1]) + p1[1]
+            sign = np.sign(p2[0] - p1[0])
+            
+        return idxs[1] * sign <= max_col_idx * sign
+
+    base_array = np.zeros(shape, dtype=dtype)  # Initialize your array of zeros
+
+    fill = np.ones(base_array.shape) * True  # Initialize boolean array defining shape fill
+
+    # Create check array for each edge segment, combine into fill array
+    for k in range(vertices.shape[0]):
+        fill = np.all([fill, check(vertices[k-1], vertices[k], base_array)], axis=0)
+    
+    print(fill.any())
+    # Set all values inside polygon to one
+    base_array[fill] = 1
+    return base_array
+
